chainregistry+lnd: moving chainregistry+chainparams to new package
